// safein/src/app/models/user.model.ts
/* ------------------------------------------------------
{
    "id": 1,
    "username": "admin",
    "password": "$2a$10$XURPShQNCsLjp1ESc2laoObo9QZDhxz73hJPaEv7/cBha4pk0AgP.",
    "email": "nomail@gmail.com",
    "nomApels": "admin",
    "role": "admin"
}
------------------------------------------------------ */


export class User {
    id?: any;
    username?: string;
    password?: string;
    email?: string;
    nomApels?: string;
    role?: string;
}